gcs/requirements.txt
PyQt5 == 5.12.0
PyQtWebEngine == 5.12.0
pymavlink
pyserial